chore : refactor trba package with triton/core modules
