+++ trba/requirements.txt
@@ -1,4 +1,5 @@
 numpy==1.18.2
+natsort==7.1.1
 Pillow==8.3.2
 pyparsing==2.4.7
 pyrsistent==0.18.0
@@ -12,3 +13,7 @@ torch==1.9.1
 torchvision==0.10.1
 tornado==6.1
 zipp==3.6.0
+lmdb==1.2.1
+tritonclient[all]==2.16.0
+attrdict==2.0.1
+torch==1.9.1
