+++ trba/src/trba/trba_triton_detector.py
@@ -1,24 +1,15 @@
 
 import json
-
-from trba.trbaOcr import TrbaOCR
 import numpy as np
-from functools import partial
-
-
-from trba.dataset import PillowImageDataset, AlignCollate
-
-
-import tritonclient.grpc as grpcclient
-import tritonclient.http as httpclient
-from tritonclient.utils import InferenceServerException
 
 import torch
-
 import torch.utils.data
 import torch.nn.functional as F
 
-from trba.trba_triton_client import TRBATritonClient
+
+from trba.core.trbaOcr import TrbaOCR
+from trba.core.dataset import PillowImageDataset, AlignCollate
+from trba.triton.trba_triton_client import TRBATritonClient
 
 class TRITON_OCR_FLAGS():
     """
@@ -40,7 +31,6 @@ class TRITON_OCR_FLAGS():
         self.streaming     = triton_flags["TRITON_OCR_STREAMING"]                                                       # Use streaming inference API. The flag is only available with gRPC protocol.
        
 
-from trba.trbaOcr import TrbaOCR
 
 class TRBATritonDetector:
